+++ app/src/main/java/com/inknironapps/bagger/data/db/dao/BagDao.kt
@@ -18,4 +18,7 @@ interface BagDao {
 
     @Query("SELECT * FROM bags WHERE id = :id")
     suspend fun getById(id: String): BagEntity?
+
+    @Query("SELECT * FROM bags")
+    suspend fun getAllOnce(): List<BagEntity>
 }